+++ cpp/NetworkServer.hpp
@@ -103,7 +103,7 @@ namespace hsd
             hsd::pair< hsd::u8sstream&, net::received_state > receive()
             {
                 _clear_buf();
-                long _response = 0;
+                isize _response = 0;
                 
                 if(_protocol == net::protocol_type::ipv4)
                 {
@@ -115,13 +115,13 @@ namespace hsd
                     _response = recvfrom(_sock.get_listening(), _net_buf.data(), 
                         4096, 0, reinterpret_cast<sockaddr*>(&_hintv6), &_len);
                 }
-                if (_response == static_cast<long>(net::received_state::err))
+                if (_response == static_cast<isize>(net::received_state::err))
                 {
                     hsd::io::err_print("Error in receiving\n");
                     _clear_buf();
                     return {_net_buf, net::received_state::err};
                 }
-                if (_response == static_cast<long>(net::received_state::disconnected))
+                if (_response == static_cast<isize>(net::received_state::disconnected))
                 {
                     hsd::io::err_print("Client disconnected\n");
                     _clear_buf();
@@ -157,7 +157,7 @@ namespace hsd
                         _send_buf += _fmt_buf[index];
                     }
 
-                    long _response = 0;
+                    isize _response = 0;
 
                     if(_protocol == net::protocol_type::ipv4)
                     {
@@ -169,7 +169,7 @@ namespace hsd
                         _response = sendto(_sock.get_listening(), _send_buf.data(), 
                             _send_buf.size(), 0, reinterpret_cast<sockaddr*>(&_hintv6), _len);
                     }
-                    if(_response == static_cast<long>(net::received_state::err))
+                    if(_response == static_cast<isize>(net::received_state::err))
                     {
                         hsd::io::err_print("Error in sending\n");
                         return net::received_state::err;
@@ -348,16 +348,16 @@ namespace hsd
             hsd::pair< hsd::u8sstream&, net::received_state > receive()
             {
                 _clear_buf();
-                long _response = recv(_sock.get_sock(), 
+                isize _response = recv(_sock.get_sock(), 
                     _net_buf.data(), 4096, 0);
 
-                if (_response == static_cast<long>(net::received_state::err))
+                if (_response == static_cast<isize>(net::received_state::err))
                 {
                     hsd::io::err_print("Error in receiving\n");
                     _clear_buf();
                     return {_net_buf, net::received_state::err};
                 }
-                if (_response == static_cast<long>(net::received_state::disconnected))
+                if (_response == static_cast<isize>(net::received_state::disconnected))
                 {
                     hsd::io::err_print("Client disconnected\n");
                     _clear_buf();
@@ -393,10 +393,10 @@ namespace hsd
                         _send_buf += _fmt_buf[index];
                     }
 
-                    long _response = send(_sock.get_sock(), 
+                    isize _response = send(_sock.get_sock(), 
                         _send_buf.data(), _send_buf.size(), 0);
 
-                    if(_response == static_cast<long>(net::received_state::err))
+                    if(_response == static_cast<isize>(net::received_state::err))
                     {
                         hsd::io::err_print("Error in sending\n");
                         return net::received_state::err;
